+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+- Display each failure iteration in rspec summary.
+
 ## [0.1.0] - 2025-07-10
 
 - Initial release